RET-2025: Refactored after review
